+++ python/coinbase-agentkit/coinbase_agentkit/action_providers/weth/constants.py
@@ -29,4 +29,4 @@ WETH_ABI = [
         "stateMutability": "view",
         "type": "function",
     },
-] 
+]
